Move jore3 NodeType under enumerated package
- NodeTypes are defined in the jr_koodisto table
