+++ src/main/java/fi/hsl/jore/importer/feature/batch/node/support/NodeTypeMapper.java
@@ -6,17 +6,17 @@ import io.vavr.collection.Map;
 
 public final class NodeTypeMapper {
 
-    public static final Map<fi.hsl.jore.importer.feature.jore3.field.NodeType, NodeType> TO_NODE_TYPE =
+    public static final Map<fi.hsl.jore.importer.feature.jore3.enumerated.NodeType, NodeType> TO_NODE_TYPE =
             HashMap.of(
-                    fi.hsl.jore.importer.feature.jore3.field.NodeType.BUS_STOP, NodeType.STOP,
-                    fi.hsl.jore.importer.feature.jore3.field.NodeType.CROSSROADS, NodeType.CROSSROADS,
-                    fi.hsl.jore.importer.feature.jore3.field.NodeType.MUNICIPAL_BORDER, NodeType.BORDER
+                    fi.hsl.jore.importer.feature.jore3.enumerated.NodeType.BUS_STOP, NodeType.STOP,
+                    fi.hsl.jore.importer.feature.jore3.enumerated.NodeType.CROSSROADS, NodeType.CROSSROADS,
+                    fi.hsl.jore.importer.feature.jore3.enumerated.NodeType.MUNICIPAL_BORDER, NodeType.BORDER
             );
 
     private NodeTypeMapper() {
     }
 
-    public static NodeType resolveNodeType(final fi.hsl.jore.importer.feature.jore3.field.NodeType nodeType) {
+    public static NodeType resolveNodeType(final fi.hsl.jore.importer.feature.jore3.enumerated.NodeType nodeType) {
         return TO_NODE_TYPE.getOrElse(nodeType,
                                       NodeType.UNKNOWN);
     }
